refactor: EditorSettings型をtypes.tsに統一し二重定義を解消
レビュー指摘対応: FlowEditor.tsxのインライン型定義と
RightPanel.tsxのexport interface EditorSettingsを
types.tsに統合し、両方からimportする形に変更。
